typedef size_t

+++ utils.c
@@ -1,7 +1,7 @@
 #include <utils.h>
 
 // NOTE: Insecure (does not detect overlapping memory)
-void memcpy(void *dest, const void *src, unsigned int n) {
+void memcpy(void *dest, const void *src, size_t n) {
     for (int i = 0; i < n; i++) {
         *(unsigned char *)dest = *(unsigned char *)src;
         ++dest;
